added mount identifier class and split up utils module
