+++ fileformats/core/__init__.py
@@ -3,9 +3,10 @@ from .classifier import Classifier
 from .datatype import DataType
 from .fileset import FileSet, MockMixin
 from .field import Field
-from .utils import (
+from .identification import (
     to_mime,
     from_mime,
     find_matching,
     from_paths,
 )
+from .sampling import SampleFileGenerator
